% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REWarDS.R
\name{definedDailyDose}
\alias{definedDailyDose}
\title{Defined daily dose}
\usage{
definedDailyDose(
  data,
  unit = 1,
  dspd_qty,
  strength,
  id,
  tot_dose_disp = NULL,
  Pt_level = FALSE
)
}
\arguments{
\item{data}{Sample simulated data}

\item{unit}{Unit of defined daily dose. Default: 1 unit = 7.5 mg}

\item{dspd_qty}{Dispensed quantity: Number of the dispensed tablets to the patient
at each prescription fill}

\item{strength}{Strength of the tablet dispensed in milligrams}

\item{id}{Unique patient identification number}

\item{tot_dose_disp}{Total dose dispensed:
dispensed quantity x strength of the tablets dispensed for
each prescription fill}

\item{Pt_level}{When TRUE, the estimated daily dose and days' supply are averaged
for the patient.}
}
\value{
DD_1_Rx_dose:  Daily dose for prescription

DD_1_Rx_DS:    Days' supply for prescription

DD_1_Pt_dose:  Average daily dose for patient

DD_1_Pt_DS:    Average days' supply for patient
}
\description{
definedDailyDose is a function that uses the World Health Organization's (WHO)
defined daily dose method to compute the daily dose and days' supply for prescriptions.
This method assumes an average daily consumption of a fixed dose, the defined daily dose
(DDD),pre-specified by WHO.
}
\details{
The DDD method can be used for any medication.
However, its accuracy has been shown to differ between drug classes.
}
\examples{
#Patient collects 100 tablets of 5 mg warfarin on January 3rd,
#and 100 tablets of 7 mg warfarin on February 1st.

## Using 1 unit DDD:
definedDailyDose(data = data, unit = 1, Pt_level = F,id = "ID",
                 dspd_qty = "DSPD_QTY", strength = "strength",
                 tot_dose_disp = NULL)

#tot_dose_disp= 500mg on January 3rd and 700 mg for February 1st.
#DDD_1_Rx_dose: 7.5 mg for each prescription fill
#DDD_1_Rx_DS is: For Jan 3rd:  500/7.5 = 66.66 day;
#                 For Feb 1st: 700/7.5=93.33 days

# pt_level can be set as TRUE to get mean values for each patient
#DDD_1_Pt_dose: (7.5+ 7.5)/2 = 7.5 mg
#DDD_1_Pt_DS: (66.66+93.33)/2 = 79.99 days


}
